feat (modal): add apply and clear buttons in modal
